kv_translate: ask whether the residual error is worth recomputing
A4, plus the storage change that lets the behavioural correction see
more than twenty prompts.

The cost model is the part of selective repair that is easy to get
wrong, so it is stated before any measurement. A target layer's native
key and value block cannot be obtained on its own: it is a function of
the residual stream at that depth, which every earlier layer produces.
Repairing layer twenty therefore means running the target's prefill
through layer twenty. Repair is a prefix, not a subset. Natively
computing the first N layers and translating the rest reintroduces N
of L layers of prefill; natively computing one late layer costs nearly
all of it. A scheme that claims to repair scattered deep layers
cheaply has miscounted, and this file is written so that mistake
cannot be made quietly.

Two measurements follow. A per-layer damage profile gives one layer
its native block and translates every other, which says where the
error lives even where fixing it would be expensive; it is diagnostic
and deliberately ignores cost. The prefix sweep is the deployable
shape and reports divergence against the fraction of prefill
reintroduced, with that fraction measured rather than assumed
proportional to depth, because attention is quadratic in context while
the rest of a layer is linear and the embedding sits outside the loop.

The verdict is reported as gap closed per unit of reintroduced compute
rather than as a quality number, since a hybrid that approaches a full
prefill is not a transfer win however good its divergence looks.

The correction cache now holds activations in half precision. These
are activations feeding a small correction, and the storage was what
capped how many distinct prompts the correction could see -- which the
sixfold train-to-held-back gap at twenty prompts says is the binding
constraint on it.

Generated-by: Claude Opus 5 (1M context)

+++ research/kv_translate/run_a3.py
@@ -263,12 +263,16 @@ def main() -> int:
                 models["target"], tp.keys_post, tp.values, cont, args.ctx
             ).float()
             ref_lp = torch.log_softmax(ref, -1).half().cpu()
+        # Held in half precision. These are activations, the correction they
+        # train is small, and the storage is what caps how many prompts the
+        # correction can see -- which the overfitting measured at twenty
+        # prompts says is the binding constraint.
         train_cache.append(
             {
-                "Xk": Xk.detach().cpu(),
-                "Xv": Xv.detach().cpu(),
-                "Yk": [k.detach().cpu() for k in to_content_keys(tp)],
-                "Yv": [v.detach().float().cpu() for v in tp.values],
+                "Xk": Xk.detach().half().cpu(),
+                "Xv": Xv.detach().half().cpu(),
+                "Yk": [k.detach().half().cpu() for k in to_content_keys(tp)],
+                "Yv": [v.detach().half().cpu() for v in tp.values],
                 "cont": cont.cpu(),
                 "ref_lp": ref_lp,
             }
@@ -289,12 +293,12 @@ def main() -> int:
         tot = []
         with torch.no_grad():
             for item in items:
-                Xk, Xv = item["Xk"].to(dev), item["Xv"].to(dev)
+                Xk, Xv = item["Xk"].to(dev).float(), item["Xv"].to(dev).float()
                 pk, pv = rk(Xk), rv(Xv)
                 if objective == "kv":
                     val = kv_error_loss(
-                        pk, [y.to(dev) for y in item["Yk"]]
-                    ) + kv_error_loss(pv, [y.to(dev) for y in item["Yv"]])
+                        pk, [y.to(dev).float() for y in item["Yk"]]
+                    ) + kv_error_loss(pv, [y.to(dev).float() for y in item["Yv"]])
                 else:
                     keys = [rerot(k.float(), pos, tg.rope_theta).to(dtype) for k in pk]
                     lg = continuation_logits(
@@ -322,13 +326,13 @@ def main() -> int:
         g = torch.Generator().manual_seed(args.seed)
         for step in range(steps):
             item = fit_cache[int(torch.randint(len(fit_cache), (1,), generator=g))]
-            Xk = item["Xk"].to(dev)
-            Xv = item["Xv"].to(dev)
+            Xk = item["Xk"].to(dev).float()
+            Xv = item["Xv"].to(dev).float()
             pk, pv = rk(Xk), rv(Xv)
             if objective == "kv":
                 loss = kv_error_loss(
-                    pk, [y.to(dev) for y in item["Yk"]]
-                ) + kv_error_loss(pv, [y.to(dev) for y in item["Yv"]])
+                    pk, [y.to(dev).float() for y in item["Yk"]]
+                ) + kv_error_loss(pv, [y.to(dev).float() for y in item["Yv"]])
             else:
                 keys = [rerot(k.float(), pos, tg.rope_theta).to(dtype) for k in pk]
                 vals = [v.to(dtype) for v in pv]
